requirements.txt
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
datadispatch==1.0.0
et-xmlfile==1.1.0
fake-useragent==1.5.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pydantic-settings==2.2.1
pytest==8.1.1
pytest-cov==5.0.0
pytest-dotenv==0.5.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tasks==2.8.0
typing_extensions==4.12.0
urllib3==2.2.0
uvicorn==0.27.1
websockets==12.0
yarl==1.9.4
